requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.3.5
aiohttp==3.10.3
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
anyio==4.8.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.7.2
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==24.2.0
babel==2.17.0
backcall==0.2.0
beautifulsoup4==4.13.3
bleach==6.2.0
cachetools==5.3.2
certifi==2024.2.2
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
comm==0.2.2
contourpy==1.2.0
cycler==0.12.1
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
Django==5.0.3
docker-pycreds==0.4.0
docopt==0.6.2
easydict==1.13
entrypoints==0.4
exceptiongroup==1.2.0
executing==2.0.1
Farama-Notifications==0.0.4
fastjsonschema==2.21.1
filelock==3.13.1
fonttools==4.48.1
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.2.0
func-timeout==4.3.5
future==1.0.0
gitdb==4.0.11
GitPython==3.1.43
glcontext==3.0.0
glfw==1.12.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
grpcio==1.60.1
gym==0.26.2
gym-notices==0.0.8
gym3==0.3.3
gymnasium==0.29.1
h11==0.14.0
h5py==3.13.0
httpcore==1.0.7
httpx==0.28.1
hydra-core==1.2.0
idna==3.6
image==1.5.33
imageio==2.37.0
imageio-ffmpeg==0.3.0
ipykernel==6.29.5
ipython==8.12.3
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.2
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kiwisolver==1.4.5
learning-action==0.0.0
learning-cv==0.0.0
learning-gazebo==0.0.0
learning-gazebo-fortress==0.0.0
learning-interface==0.0.0
learning-launch==0.0.0
learning-node==0.0.0
learning-parameter==0.0.0
learning-pkg-python==0.0.0
learning-qos==0.0.0
learning-service==0.0.0
learning-tf==0.0.0
learning-topic==0.0.0
learning-urdf==0.0.0
lightning-utilities==0.11.6
llvmlite==0.44.0
lmdb==1.6.2
Markdown==3.5.2
MarkupSafe==2.1.5
matplotlib==3.8.2
matplotlib-inline==0.1.6
minigrid==2.3.1
mistune==3.1.1
moderngl==5.12.0
mpmath==1.3.0
msgpack==1.1.0
msgpack-numpy==0.4.8
multidict==6.0.5
narwhals==1.26.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest_asyncio==1.6.0
networkx==3.2.1
nn==0.1.1
notebook==7.3.2
notebook_shim==0.2.4
numba==0.61.2
numpy==1.26.4
numpy-groupies==0.11.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.3.0
overrides==7.7.0
packaging==23.2
pandas==2.2.0
pandocfilters==1.5.1
parso==0.8.3
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.2.0
pip==24.0
pipreqs==0.5.0
platformdirs==4.3.6
plotly==6.0.0
procgen==0.10.7
prometheus_client==0.21.1
prompt-toolkit==3.0.43
protobuf==3.20.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.22
pyDeprecate==0.3.2
pygame==2.5.2
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0
python-dotenv==0.20.0
python-json-logger==3.2.1
pytorch-lightning==1.6.3
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.36.2
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
ribs==0.7.1
rpds-py==0.22.3
rsa==4.9
scikit-learn==1.6.1
scipy==1.12.0
seaborn==0.13.2
Send2Trash==1.8.3
sentry-sdk==1.44.1
setproctitle==1.3.3
setuptools==68.2.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.4.4
stable-baselines3==2.2.1
stack-data==0.6.3
sympy==1.12
tensorboard==2.19.0
tensorboard-data-server==0.7.2
terminado==0.18.1
thop==0.1.1.post2209072238
threadpoolctl==3.6.0
tinycss2==1.4.0
tomli==2.2.1
torch==2.2.1
torch-lr-finder==0.2.1
torchmetrics==1.4.1
torchvision==0.17.1
tornado==6.4.2
tqdm==4.66.1
traitlets==5.14.1
triton==2.2.0
types-python-dateutil==2.9.0.20241206
typing_extensions==4.10.0
tzdata==2023.4
uri-template==1.3.0
urllib3==2.2.1
wandb==0.16.6
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
wheel==0.41.2
yarg==0.1.9
yarl==1.9.4
